@@ -1,36 +1,39 @@
 {
-  "name": "Chat",
+  "name": "ChatDemo",
   "version": "0.0.1",
   "private": true,
   "scripts": {
+    "android": "react-native run-android",
+    "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint .",
-    "ios": "react-native run-ios",
-    "android": "react-native run-android"
+    "lint": "eslint ."
   },
   "dependencies": {
-    "@react-native-community/masked-view": "^0.1.6",
-    "react": "16.9.0",
-    "react-native": "0.61.5",
-    "react-native-gesture-handler": "^1.5.3",
-    "react-native-reanimated": "^1.7.0",
-    "react-native-safe-area-context": "^0.6.2",
-    "react-native-screens": "^2.0.0-alpha.27",
-    "react-navigation": "^4.0.10",
-    "react-navigation-stack": "^2.0.15",
-    "unique-names-generator": "^4.1.1"
+    "@react-native-community/masked-view": "^0.1.10",
+    "@react-native-firebase/app": "^8.4.5",
+    "@react-native-firebase/auth": "^9.3.0",
+    "@react-native-firebase/firestore": "^7.8.7",
+    "@react-navigation/native": "^5.7.6",
+    "@react-navigation/stack": "^5.9.3",
+    "react": "16.13.1",
+    "react-native": "0.63.3",
+    "react-native-gesture-handler": "^1.8.0",
+    "react-native-gifted-chat": "^0.16.3",
+    "react-native-reanimated": "^1.13.1",
+    "react-native-safe-area-context": "^3.1.8",
+    "react-native-screens": "^2.11.0",
+    "unique-names-generator": "^4.3.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.2",
-    "@babel/runtime": "^7.6.2",
-    "babel-jest": "^24.9.0",
-    "eslint": "^6.8.0",
-    "eslint-config-handlebarlabs": "^0.0.4",
-    "jest": "^24.9.0",
-    "metro-react-native-babel-preset": "^0.56.0",
-    "prettier": "^1.19.1",
-    "react-test-renderer": "16.9.0"
+    "@babel/core": "^7.11.6",
+    "@babel/runtime": "^7.11.2",
+    "babel-jest": "^26.5.2",
+    "eslint": "^7.11.0",
+    "eslint-config-handlebarlabs": "^0.0.6",
+    "jest": "^26.5.3",
+    "metro-react-native-babel-preset": "^0.63.0",
+    "react-test-renderer": "16.13.1"
   },
   "jest": {
     "preset": "react-native"
