@@ -1,46 +1,68 @@
-import React from 'react';
-import {FlatList} from 'react-native';
+import React, { useState, useEffect } from 'react';
+import { FlatList, View } from 'react-native';
 
-import {ThreadRow, Separator} from '../components/ThreadRow';
+import { ThreadRow, Separator } from '../components/ThreadRow';
+import { listenToThreads, listenToThreadTracking } from '../firebase';
 
-export default class Threads extends React.Component {
-  state = {
-    threads: [
-      {
-        _id: '1',
-        unread: false,
-        name: 'Demo 1',
-        latestMessage: {
-          text: 'Hello, this is an example.',
-        },
-      },
-      {
-        _id: '2',
-        unread: true,
-        name: 'Demo 2',
-        latestMessage: {
-          text: 'Hello, this is another example.',
-        },
-      },
-    ],
-  };
+const isThreadUnread = (thread, threadTracking) => {
+  if (
+    threadTracking[thread._id] &&
+    threadTracking[thread._id] < thread.latestMessage.createdAt
+  ) {
+    return true;
+  }
 
-  render() {
-    return (
+  return false;
+};
+
+export default ({ navigation }) => {
+  const [threads, setThreads] = useState([]);
+  const [threadTracking, setThreadTracking] = useState({});
+
+  useEffect(() => {
+    const unsubscribe = listenToThreads().onSnapshot((querySnapshot) => {
+      // console.log(querySnapshot.docs);
+      const allThreads = querySnapshot.docs.map((snapshot) => {
+        return {
+          _id: snapshot.id,
+          name: '',
+          latestMessage: { text: '' },
+          ...snapshot.data(),
+        };
+      });
+
+      setThreads(allThreads);
+    });
+
+    return () => {
+      unsubscribe();
+    };
+  }, []);
+
+  useEffect(() => {
+    const unsubscribe = listenToThreadTracking().onSnapshot((snapshot) => {
+      setThreadTracking(snapshot.data() || {});
+    });
+
+    return () => {
+      unsubscribe();
+    };
+  }, []);
+
+  return (
+    <View style={{ flex: 1, backgroundColor: '#fff', paddingBottom: 50 }}>
       <FlatList
-        data={this.state.threads}
-        keyExtractor={item => item._id}
-        renderItem={({item}) => (
+        data={threads}
+        keyExtractor={(item) => item._id}
+        renderItem={({ item }) => (
           <ThreadRow
             {...item}
-            onPress={() =>
-              this.props.navigation.navigate('Messages', {thread: item})
-            }
-            unread={item.unread}
+            onPress={() => navigation.navigate('Messages', { thread: item })}
+            unread={isThreadUnread(item, threadTracking)}
           />
         )}
         ItemSeparatorComponent={() => <Separator />}
       />
-    );
-  }
-}
+    </View>
+  );
+};
